CCT 1.0.18
Comprehensive updates to how zooming is handled in all Waveform plots to support ‘sticky’ zoom levels
Additional toggle included to allow zooming an entire group of Waveform plots to match an index plot’s zoom
Additional updates to uncertainty quantification metrics based on continuing iteration with the user community
Include a coda start offset parameter in the frequency band table to allow measurement offsets for entire frequency bands; 'CS' picks
Implementation of user adjustable start times for coda measurements on an individual waveform basis; 'UCS' picks, overrides 'CS' values, if present
Include reference features in several plots to allow comparison against known terms and measurements.
Basic tutorial and notes for new users and basic API documentation for developers now included in the help menu
Changed the way coda start is calculated to make it a two-step where we find the max vel, then fit the model, then back-prop the expected velocity into the coda start time field.
Fix for zoom issues when you go straight to measurement/site and box select without displaying any single waveform first.
Bump to TLS 1.3 since support has been backported to Java 8
Changed the map screenshot button to put the icons back onto png
Moving picks on waveforms with site measurements will re-measure the measurement and replot it in site automatically.
Dynamically update synthetic when you move UCS pick.
New top axis option for the Moment vs Energy plot + other updates related.
Fixed log_x axis resizing not being taken into acount for the station correction sd plot.
New REST endpoint to directly generate a synthetic coda envelope based on the loaded calibration.
Fixed a bug where if you had ref/val events loaded not in data set the site tab could blow up a bit.
Updates to default table values to increase the default window lengths when no tables are loaded, based on user feedback.

+++ calibration-gui/src/main/java/gov/llnl/gnem/apps/coda/common/gui/converters/sac/SacExporter.java
@@ -81,15 +81,15 @@ public class SacExporter {
                     try {
                         WaveformPick pick = w.getAssociatedPicks().get(i);
                         if (pick.getPickType().equalsIgnoreCase(PICK_TYPES.F.getPhase())) {
-                            header.f = pick.getPickTimeSecFromOrigin();
+                            header.f = pick.getPickTimeSecFromOrigin().floatValue();
                         } else if (pick.getPickType().equalsIgnoreCase(PICK_TYPES.A.getPhase())) {
-                            header.a = pick.getPickTimeSecFromOrigin();
+                            header.a = pick.getPickTimeSecFromOrigin().floatValue();
                         } else if (pick.getPickType().equalsIgnoreCase(PICK_TYPES.B.getPhase())) {
-                            header.b = pick.getPickTimeSecFromOrigin();
+                            header.b = pick.getPickTimeSecFromOrigin().floatValue();
                         } else if (pick.getPickType().equalsIgnoreCase(PICK_TYPES.O.getPhase())) {
-                            header.o = pick.getPickTimeSecFromOrigin();
+                            header.o = pick.getPickTimeSecFromOrigin().floatValue();
                         } else {
-                            header.setTimePick(i, pick.getPickTimeSecFromOrigin(), pick.getPickName());
+                            header.setTimePick(i, pick.getPickTimeSecFromOrigin().floatValue(), pick.getPickName());
                         }
                     } catch (RuntimeException e) {
                         log.error(e.getMessage(), e);
