+++ calibration-service/calibration-service-impl/src/main/java/gov/llnl/gnem/apps/coda/calibration/service/impl/ShapeCalibrationServiceImpl.java
@@ -151,7 +151,7 @@ public class ShapeCalibrationServiceImpl implements ShapeCalibrationService {
                     //Ensure pick is propagated back to waveform
                     double end = curve.getEndTime();
                     if (end >= 0d) {
-                        endPick.setPickTimeSecFromOrigin((float) end);
+                        endPick.setPickTimeSecFromOrigin(end);
                     }
                 }
             }
